+++ Mailer/extensions/mailer-module/src/Tests/Unit/ContentGenerator/Replace/RtmReplaceTest.php
@@ -116,4 +116,19 @@ class RtmReplaceTest extends TestCase
         $url = $this->rtmReplace->replaceUrl('https://predplatne.dennikn.sk/email-settings?%recipient.autologin%', $this->generatorInput);
         $this->assertEquals('https://predplatne.dennikn.sk/email-settings?%recipient.autologin%&rtm_source=demo-weekly-newsletter&rtm_medium=email&rtm_campaign=impresa_mail_20190903103350&rtm_content=', $url);
     }
+
+    public function testMultilineAnchorDefinition(): void
+    {
+        $content = $this->rtmReplace->replace('
+            <a href="https://expresso.pt/html/que-nao-entrou-em-acao?rtm_content=apple&amp;feira=terca&modelo=1%3Fa%3D1&b=2" target="blank" style="color:#ffffff;
+                display:inline-block;"
+            />',
+            $this->generatorInput);
+        $this->assertEquals('
+            <a href="https://expresso.pt/html/que-nao-entrou-em-acao?rtm_content=&amp;feira=terca&modelo=1%3Fa%3D1&b=2&rtm_source=demo-weekly-newsletter&rtm_medium=email&rtm_campaign=impresa_mail_20190903103350" target="blank" style="color:#ffffff;
+                display:inline-block;"
+            />',
+            $content
+        );
+    }
 }